moved submitassignment to stu
