+++ app/stu/[courseCode]/submitAssignment/[name]/page.jsx
@@ -1,32 +1,47 @@
 'use client'
-import CourseNavBar from '../../components/CourseNavBar';
-import Sidebar from '../../components/Sidebar';
+import Sidebar from '@/app/components/Sidebar';
+import CourseNavBar from '@/app/components/CourseNavBar';
 import { useEffect, useState } from 'react';
 import { onAuthStateChanged } from 'firebase/auth';
 import { useParams } from 'next/navigation';
 import { auth } from '@/app/lib/firebase';
 import { getDoc, doc,getDocs,query,collection, where } from 'firebase/firestore';
-import db from '../../lib/firebase';
-import { useSearchParams } from 'next/navigation';
+import db from '../../../../lib/firebase';
 
 
 
 export default function Assignments() {
-    console.log(useParams());
-
+    const {name,courseCode} = useParams();
+    console.log(name);
     // const search = window.location.search;
     // const params = new URLSearchParams(search);
     // console.log(params);
 
     const [currentAssignments, setCurrentAssignments] = useState([]);
     const [user,setUser] = useState(null);
-    const [userType,setUserType] = useState('user');
 
     useEffect(() => {
         const unsubscribe = onAuthStateChanged(auth, async (user) => {
             if (auth.currentUser) {
-               
-              
+                setUser(auth.currentUser);
+
+                const assignmentRef = doc(db,'quizzes',name);
+                const assignmentSnapshot = await getDoc(assignmentRef);
+
+                if(!assignmentSnapshot.empty){
+                    console.log(assignmentSnapshot.data())
+                }else{
+                    const assignmentRef = doc(db,'essays',name);
+                    const assignmentSnapshot = await getDoc(assignmentRef);
+                    
+                    if(!assignmentSnapshot.empty){
+                        console.log(assignmentSnapshot.data());
+                    }
+
+
+                }
+
+                
                   
             }
         });
@@ -38,7 +53,7 @@ export default function Assignments() {
         <div className="flex flex-col md:flex-row bg-blue-100">
             <Sidebar />
             <div className="relative md:ml-64">
-                <CourseNavBar />
+                <CourseNavBar courseCode={courseCode} />
             </div>
             <div className="p-6 text-center w-full">
                 <h1 className="text-3xl text-black font-semibold mb-4" data-testid="course-heading">Course Name</h1>
@@ -52,4 +67,5 @@ export default function Assignments() {
             </div>
         </div>
     );
+
 }
